split out helpers and initialization to a separate file
